decoupling db from string and clsuter
