+++ oagdedupe/fastapi/fapi.py
@@ -102,7 +102,7 @@ class TasksPost(SettingsEnabler):
         """
 
         logging.info("getting distances")
-        distances = self.api.orm.get_distances()
+        distances = self.api.compute.get_distances()
 
         logging.info("getting active learning samples")
         sample_idx, _ = self.clf.query(
@@ -110,7 +110,7 @@ class TasksPost(SettingsEnabler):
             n_instances=n_instances,
         )
 
-        return distances.loc[sample_idx, self.api.orm.compare_cols]
+        return distances.loc[sample_idx, self.api.compute.compare_cols]
 
     def _post_tasks(self) -> None:
         """
@@ -131,10 +131,10 @@ class TasksGet(SettingsEnabler):
         set "labelled" column in train table to True
         """
         indices = set(newlabels["_index_l"]).union(set(newlabels["_index_r"]))
-        with self.api.init.Session() as session:
+        with self.api.compute.Session() as session:
             stmt = (
-                update(self.api.init.Train)
-                .where(self.api.init.Train._index.in_(indices))
+                update(self.api.compute.Train)
+                .where(self.api.compute.Train._index.in_(indices))
                 .values(labelled=True)
             )
             session.execute(stmt)
@@ -144,7 +144,7 @@ class TasksGet(SettingsEnabler):
         """
         add new labels to labels table
         """
-        self.api.orm._update_table(newlabels, self.api.init.Labels())
+        self.api.compute._update_table(newlabels, self.api.compute.Labels())
 
     def _get_new_labels(self) -> pd.DataFrame:
         """
@@ -185,7 +185,7 @@ class Model(TasksGet, TasksPost, Projects):
         """
         trains active learning model
         """
-        labels = self.api.orm.get_labels()
+        labels = self.api.compute.get_labels()
         self.clf.teach(
             labels[self.settings.attributes].values,
             labels["label"].values,
